table_process/extract_tables/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
python-docx>=0.8.11
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
PyYAML>=5.4.0
openpyxl>=3.0.0
